requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
Flask-Login==0.6.3
pandas==2.2.2
pytest==8.2.0
